Extract layout components into their own module, and then tighten up import rules further
