+++ src/modules/layout/index.js
@@ -1,0 +1,4 @@
+export { default as NotFoundPage } from "./pages/NotFoundPage";
+export { default as DashboardPage } from "./pages/DashboardPage";
+export { default as Page } from "./components/Page";
+export { default as AppFrame } from "./components/AppFrame";
