+++ src/modules/layout/pages/NotFoundPage.js
@@ -4,8 +4,7 @@ import { withStyles } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import Paper from "@material-ui/core/Paper";
 
-import Page from "components/Page";
-import AppFrame from "components/AppFrame";
+import { Page, AppFrame } from "modules/layout";
 
 const styles = theme => ({
   div: {
