+++ src/modules/signup/index.js
@@ -1,0 +1,6 @@
+export { default as InvitePage } from "./pages/InvitePage";
+export {
+  default as CreateOrganizationPage
+} from "./pages/CreateOrganizationPage";
+export { default as JoinPage } from "./pages/JoinPage";
+export { default as HomePage } from "./pages/HomePage";
