+++ src/modules/signup/pages/JoinPage.js
@@ -4,8 +4,8 @@ import { FirestoreDocument } from "react-firestore";
 import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
 
+import { Page } from "modules/layout";
 import Progress from "components/Progress.js";
-import Page from "components/Page";
 import { createUserAndProfile } from "auth";
 import { handleError } from "utils";
